requirements.txt
streamlit==1.22.0
yt-dlp==2023.3.4
openai==0.27.0
librosa==0.10.0
numpy==1.24.3
matplotlib==3.7.1
python-dotenv==1.0.0
